+++ fcp/src/cmd_trunc.c
@@ -77,12 +77,12 @@ static int __trunc(args_t * args, off_t offset)
 	if (ffs->count <= 0)
 		return 0;
 
-	size_t block_size;
+	uint32_t block_size;
 	if (__ffs_info(ffs, FFS_INFO_BLOCK_SIZE, &block_size) < 0)
 		return -1;
 
 	if (args->buffer != NULL) {
-		size_t buffer;
+		uint32_t buffer;
 		if (parse_size(args->buffer, &buffer) < 0)
 			return -1;
 		if (__ffs_buffer(ffs, buffer) < 0)
@@ -103,7 +103,7 @@ static int __trunc(args_t * args, off_t offset)
 	if (entry.type == FFS_TYPE_LOGICAL) {
 		if (args->verbose == f_VERBOSE)
 			fprintf(stderr, "%8llx: %s: logical (skip)\n",
-				offset, full_name);
+				(long long)offset, full_name);
 		return 0;
 	}
 
@@ -111,11 +111,11 @@ static int __trunc(args_t * args, off_t offset)
 	    entry.flags && FFS_FLAGS_PROTECTED) {
 		if (args->verbose == f_VERBOSE)
 			fprintf(stderr, "%8llx: %s: protected (skip)\n",
-				offset, full_name);
+				(long long)offset, full_name);
 		return 0;
 	}
 
-	size_t size = 0;
+	uint32_t size = 0;
 	if (args->opt_nr == 1) {
 		size = entry.size * block_size;
 	} else if (args->opt_nr == 2) {
@@ -130,7 +130,7 @@ static int __trunc(args_t * args, off_t offset)
 
 	if (args->verbose == f_VERBOSE)
 		fprintf(stderr, "%8llx: %s: truncate '%x' (done)\n",
-			offset, full_name, size);
+			(long long)offset, full_name, size);
 
 	return 0;
 }
